+++ zoia_lib/backend/split_master.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+i=0
+for f in *; do
+  d=dir_$(printf % 03d $((i / 64 + 1)))
+  mkdir -p "$d"
+  mv "$f" "$d"
+  ((i++))
+done
