--- zoia_lib/split_master.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-i=0;
-for f in *;
-do
-d=dir_$(printf % 03d $((i / 64+1)));
-mkdir -p "$d";
-mv "$f" "$d";
-((i++));
-done
